correctly load defaults with argparse - unified handling of --configFile - specify default values only in the parsing, nowhere else (but only for "train" and "predict", for the other modes, this still needs to be done) - put modes into separate files (train, predict, trainggn, predictgnn, interpretgnn, convert) - parse JSON configs just like CLI args (this will give the user feedback concerning syntax errors etc in the JSON, just like with CLI args) - removed jsonpickle dependency - removed py/object from JSON configs (not needed anymore) - removed aeOptimizer argument (wasn't being used) - added parsing for --fnnType - split train/predict Options into three classes (common, train, predict) - added todo's regarding some confusing parts
