+++ dfpl/callbacks.py
@@ -4,10 +4,11 @@ from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint
 # for testing in Weights & Biases
 from wandb.keras import WandbCallback
 
-from dfpl import options, settings
+from dfpl import settings
+from dfpl.train import TrainOptions
 
 
-def autoencoder_callback(checkpoint_path: str, opts: options.Options) -> list:
+def autoencoder_callback(checkpoint_path: str, opts: TrainOptions) -> list:
     """
     Callbacks for fitting the autoencoder
 
@@ -22,25 +23,15 @@ def autoencoder_callback(checkpoint_path: str, opts: options.Options) -> list:
     else:
         target = "loss"
         # enable this checkpoint to restore the weights of the best performing model
-    if opts.aeType == "deterministic":
-        checkpoint = ModelCheckpoint(
-            checkpoint_path,
-            monitor=target,
-            mode="min",
-            verbose=1,
-            save_freq="epoch",
-            save_best_only=True,
-        )
-    else:
-        checkpoint = ModelCheckpoint(
-            checkpoint_path,
-            monitor=target,
-            mode="min",
-            verbose=1,
-            save_freq="epoch",
-            save_best_only=True,
-            save_weights_only=True,
-        )
+    checkpoint = ModelCheckpoint(
+        checkpoint_path,
+        monitor=target,
+        mode="min",
+        verbose=1,
+        period=settings.ac_train_check_period,
+        save_best_only=True,
+        save_weights_only=True,
+    )
     callbacks.append(checkpoint)
 
     # enable early stopping if val_loss is not improving anymore
@@ -53,12 +44,13 @@ def autoencoder_callback(checkpoint_path: str, opts: options.Options) -> list:
         restore_best_weights=True,
     )
     callbacks.append(early_stop)
+
     if opts.aeWabTracking and not opts.wabTracking:
         callbacks.append(WandbCallback(save_model=False))
     return callbacks
 
 
-def nn_callback(checkpoint_path: str, opts: options.Options) -> list:
+def nn_callback(checkpoint_path: str, opts: TrainOptions) -> list:
     """
     Callbacks for fitting the feed forward network (FNN)
 
@@ -74,7 +66,7 @@ def nn_callback(checkpoint_path: str, opts: options.Options) -> list:
         checkpoint = ModelCheckpoint(
             checkpoint_path,
             verbose=1,
-            save_freq="epoch",
+            period=settings.nn_train_check_period,
             save_best_only=True,
             monitor="val_loss",
             mode="min",
