+++ dfpl/predictions.py
@@ -4,10 +4,12 @@ import numpy as np
 import pandas as pd
 import tensorflow.keras.models
 
-from dfpl import options, settings
+import dfpl.predict
+import dfpl.train
+from dfpl import settings
 
 
-def predict_values(df: pd.DataFrame, opts: options.Options) -> pd.DataFrame:
+def predict_values(df: pd.DataFrame, opts: dfpl.predict.PredictOptions) -> pd.DataFrame:
     """
     Predict a set of chemicals using a selected model.
 
@@ -16,7 +18,7 @@ def predict_values(df: pd.DataFrame, opts: options.Options) -> pd.DataFrame:
     :return:
     """
     model = tensorflow.keras.models.load_model(opts.fnnModelDir, compile=False)
-    model.compile(loss=opts.lossFunction, optimizer=opts.optimizer)
+    model.compile(loss=opts.lossFunction, optimizer=opts.optimizer)  # todo: Why are these parameters needed here?
     if opts.compressFeatures:
         sub_df = df[df["fpcompressed"].notnull()]
         x = np.array(
@@ -25,7 +27,7 @@ def predict_values(df: pd.DataFrame, opts: options.Options) -> pd.DataFrame:
             copy=settings.numpy_copy_values,
         )
         logging.info(f"Compressed FP matrix with shape {x.shape} and type {x.dtype}")
-        sub_df["predicted"] = pd.DataFrame(model.predict(x), columns=["predicted"])
+        sub_df["predicted"] = pd.DataFrame(dfpl.predict.predict(x), columns=["predicted"])
         return sub_df
     else:
         sub_df = df[df["fp"].notnull()]
@@ -35,5 +37,5 @@ def predict_values(df: pd.DataFrame, opts: options.Options) -> pd.DataFrame:
             copy=settings.numpy_copy_values,
         )
         logging.info(f"Uncompressed FP matrix with shape {x.shape} and type {x.dtype}")
-        sub_df["predicted"] = pd.DataFrame(model.predict(x), columns=["predicted"])
+        sub_df["predicted"] = pd.DataFrame(dfpl.predict.predict(x), columns=["predicted"])
         return sub_df
