+++ dfpl/vae.py
@@ -14,15 +14,16 @@ from tensorflow.keras.layers import Dense, Input, Lambda
 from tensorflow.keras.models import Model
 from tensorflow.python.framework.ops import disable_eager_execution
 
+from dfpl.train import TrainOptions
 from dfpl import callbacks
 from dfpl import history as ht
-from dfpl import options, settings
+from dfpl import settings
 from dfpl.utils import ae_scaffold_split, weight_split
 
 disable_eager_execution()
 
 
-def define_vae_model(opts: options.Options, output_bias=None) -> Tuple[Model, Model]:
+def define_vae_model(opts: TrainOptions, output_bias=None) -> Tuple[Model, Model]:
     input_size = opts.fpSize
     encoding_dim = (
         opts.encFPSize
@@ -154,12 +155,12 @@ def define_vae_model(opts: options.Options, output_bias=None) -> Tuple[Model, Mo
     return autoencoder, encoder
 
 
-def train_full_vae(df: pd.DataFrame, opts: options.Options) -> Model:
+def train_full_vae(df: pd.DataFrame, opts: TrainOptions) -> Model:
     """
     Trains an autoencoder on the given feature matrix X. The response matrix is only used to
     split the data into meaningful test and train sets.
 
-    :param opts: Command line arguments as defined in options.py
+    :param opts: Command line arguments
     :param df: Pandas dataframe that contains the SMILES/InChI data for training the autoencoder
     :return: The encoder model of the trained autoencoder
     """
