+++ example/train.json
@@ -1,5 +1,4 @@
 {
-  "py/object": "dfpl.options.Options",
   "inputFile": "tests/data/S_dataset.csv",
   "outputDir": "example/results_train/",
   "ecModelDir": "example/results_train/",
@@ -14,7 +13,6 @@
   "aeSplitType": "random",
   "aeEpochs": 2,
   "aeBatchSize": 351,
-  "aeOptimizer": "Adam",
   "aeActivationFunction": "relu",
   "aeLearningRate": 0.001,
   "aeLearningRateDecay": 0.96,
